requirements.txt
amocrm-api==2.6.1
cachetools==5.5.0
certifi==2024.12.14
charset-normalizer==3.4.0

gspread==6.1.4
httplib2==0.22.0
idna==3.10
oauthlib==3.2.2
proto-plus==1.25.0
protobuf==5.29.2
pyasn1==0.6.1
pyasn1_modules==0.4.1
PyJWT==2.10.1
pyparsing==3.2.0
pyTelegramBotAPI==4.25.0
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
telebot==0.0.5
uritemplate==4.1.1
urllib3==2.3.0
